fix: 상수화 및 함수 분리 완료
