+++ src/components/create/NavigationButtons.tsx
@@ -1,45 +1,18 @@
 import useLatLng from '@/hooks/map/useLatLng';
-import useMap from '@/hooks/navigation/useMap';
+import useOpenMapHandlers from '@/hooks/navigation/useNavigationButtons';
 
 const NavigationButtons = ({ address, name }: { address: string; name: string }) => {
   const coords = useLatLng(address);
-  const openMap = useMap();
+  const { openKakaoMap, openNaverMap, openTMap } = useOpenMapHandlers(
+    { lat: coords.lat, lng: coords.lng },
+    name
+  );
 
   return (
     <div>
-      <button
-        onClick={() =>
-          openMap({
-            endCoords: { epLat: coords.lat, epLng: coords.lng },
-            name: name,
-            app: 'kakao',
-          })
-        }
-      >
-        카카오맵
-      </button>
-      <button
-        onClick={() =>
-          openMap({
-            endCoords: { epLat: coords.lat, epLng: coords.lng },
-            name: name,
-            app: 'naver',
-          })
-        }
-      >
-        네이버 지도
-      </button>
-      <button
-        onClick={() =>
-          openMap({
-            endCoords: { epLat: coords.lat, epLng: coords.lng },
-            name: name,
-            app: 'tmap',
-          })
-        }
-      >
-        티맵
-      </button>
+      <button onClick={openKakaoMap}>카카오맵</button>
+      <button onClick={openNaverMap}>네이버 지도</button>
+      <button onClick={openTMap}>티맵</button>
     </div>
   );
 };
